Using feature extraction with SVM in order to classify and remove background automobile sounds from the raw audio data
Project done with Lillian Anderson at Georgia Tech for ECE 4271 in Spring 2020.
The same technique can be used to remove conversational noise, however MFCC can help improve the quality of classification for human conversation. Better methods of doing this include the use of Librosa and OpenSmile in Python along with the use of a convolutional neural net, or PCA for higher dimensions of data.